Adding 4 directional blend shaders
Adding 4 directional blend shaders, blending based on the vertex normal. Lerp_1 is simply the dot product of the up vector and the vertex normals. Lerp_2 adds controls for an offset and blending, while Lerp 3 remaps the controls to work more intuitively between 0-1. Lerp_4 adds a tileable perlin noise texture to the blend to make in more naturalistic.